finish simple project without spring security,use only session

+++ src/main/java/np/schoolmanagementsystem/service/Impl/StudentServiceImpl.java
@@ -5,7 +5,9 @@ import lombok.Data;
 import np.schoolmanagementsystem.Enum.Role;
 import np.schoolmanagementsystem.Mapper.StudentMapper;
 import np.schoolmanagementsystem.dto.StudentDto;
+import np.schoolmanagementsystem.entity.Classroom;
 import np.schoolmanagementsystem.entity.Student;
+import np.schoolmanagementsystem.repository.ClassroomRepository;
 import np.schoolmanagementsystem.repository.StudentRepository;
 import np.schoolmanagementsystem.service.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,11 +23,13 @@ import java.util.stream.Collectors;
 
 public class StudentServiceImpl implements StudentService {
 
-   private  StudentRepository studentRepository;
+    private final ClassroomRepository classroomRepository;
+    private StudentRepository studentRepository;
 
-@Autowired
-    public StudentServiceImpl(StudentRepository studentRepository) {
+    @Autowired
+    public StudentServiceImpl(StudentRepository studentRepository, ClassroomRepository classroomRepository) {
         this.studentRepository = studentRepository;
+        this.classroomRepository = classroomRepository;
     }
 
 
@@ -39,78 +43,75 @@ public class StudentServiceImpl implements StudentService {
     @Override
     public StudentDto updateStudent(StudentDto studentDto, Long studentId, HttpSession session) {
 
-//    its for session and role divide
-//        String roleName = (String) session.getAttribute("role");
-//        System.out.println(roleName);
-//        Role role = Role.valueOf(roleName);
+        Student student = StudentMapper.mapToStudent(studentDto);
+        Student updateStudent = studentRepository.save(student);
+        return StudentMapper.mapToStudentDto(updateStudent);
 
-//        if (Role.ADMIN.equals(role))
-        {
-            Student student=StudentMapper.mapToStudent(studentDto);
-            Student updateStudent=studentRepository.save(student);
-            return StudentMapper.mapToStudentDto(updateStudent);
-        }
-//        else {
-//            throw new RuntimeException("Only Admin can update student.");
-//        }
     }
 
     @Override
     public StudentDto getStudentById(Long studentId) {
-        Student student=studentRepository.findById(studentId)
-                .orElseThrow(()-> new RuntimeException("Student not found"));
+        Student student = studentRepository.findById(studentId)
+                .orElseThrow(() -> new RuntimeException("Student not found"));
 
         return StudentMapper.mapToStudentDto(student);
     }
 
     @Override
-    public StudentDto deleteStudentById(Long id) {
-        Student student=studentRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("Student not found"));
-        studentRepository.delete(student);
+    public StudentDto deleteStudentById(Long id, HttpSession session) {
 
-        return StudentMapper.mapToStudentDto(student);
+        String rollName = (String) session.getAttribute("Role");
+        if (Role.ADMIN.name().equals(rollName)) {
+            Student student = studentRepository.findById(id)
+                    .orElseThrow(() -> new RuntimeException("Student not found"));
+            studentRepository.delete(student);
+
+            return StudentMapper.mapToStudentDto(student);
+        } else {
+            throw new RuntimeException("Only Admin can delete student.");
+        }
     }
 
     @Override
     public List<StudentDto> getAllStudents() {
-    List<Student> students=studentRepository.findAll();
-    return students.stream().map(student ->StudentMapper.mapToStudentDto(student))
-            .collect(Collectors.toList());
-}
+        List<Student> students = studentRepository.findAll();
+        return students.stream().map(student -> StudentMapper.mapToStudentDto(student))
+                .collect(Collectors.toList());
+    }
 
     @Override
     public StudentDto studentRegistration(StudentDto studentDto) {
 
-            Optional<Student> existingStudent = studentRepository.findByEmail(studentDto.getEmail());
-            if (existingStudent.isPresent()) {
-                throw new RuntimeException("Student already exists");
-            }
+        Optional<Student> existingStudent = studentRepository.findByEmail(studentDto.getEmail());
+        if (existingStudent.isPresent()) {
+            throw new RuntimeException("Student already exists");
+        }
+//        Classroom classroom = classroomRepository.findById(studentDto.getclassroomId())
+//                .orElseThrow(()->new RuntimeException("classroom not found"));
 
 
-        Student student=StudentMapper.mapToStudent(studentDto);
-            student.setRole(Role.STUDENT);
-        Student savedStudent=studentRepository.save(student);
+        Student student = StudentMapper.mapToStudent(studentDto);
+        student.setRole(Role.STUDENT);
+        Student savedStudent = studentRepository.save(student);
         return StudentMapper.mapToStudentDto(savedStudent);
 
     }
 
     @Override
-    public  boolean studentLogin(String userName, String password) {
+    public boolean studentLogin(String userName, String password) {
         Optional<Student> StudentOptional = studentRepository
                 .findByUserName(userName);
 
         if (StudentOptional.isEmpty()) {
             throw new RuntimeException("Student not found ");
-        }
-        else {
+        } else {
             Student student = StudentOptional.get();
             if (!student.getUserName().equals(userName)) {
                 throw new RuntimeException("Invalid userName");
             }
 
 
-            if(!student.getPassword().equals(password)) {
+            if (!student.getPassword().equals(password)) {
                 throw new RuntimeException("Invalid password");
             }
 //            return StudentMapper.mapToStudentDto(student);
